refactor(queues): break email import cycle + drop intent non-null assertion
- Extract queue-dependent sendEmail into transport.producer.ts so the
  email queue no longer imports a module that imports it back (madge cycle gone)
- Replace defaultIndexer! non-null assertion with an IIFE-captured instance
