+++ backend/src/lib/email/tests/email-handlers.spec.ts
@@ -3,12 +3,12 @@ config({ path: '.env.test', override: true });
 
 import { describe, it, expect, jest, beforeEach, mock, afterAll } from 'bun:test';
 import { sendConnectionRequestEmail, sendConnectionAcceptedEmail } from '../notification.sender';
-import * as emailModule from '../transport.helper';
+import * as emailModule from '../transport.producer';
 import * as templatesModule from '../templates/connection-request.template'; // We need to mock specific templates now
 import * as connectionAcceptedTemplateModule from '../templates/connection-accepted.template';
 
 // Mock dependencies
-mock.module('../transport.helper', () => ({
+mock.module('../transport.producer', () => ({
   sendEmail: jest.fn()
 }));
